Continue HR management system development

+++ practicum/BackEnd/HR System/DTOs/Dashboard/DashboardIssueDto.cs
@@ -1,7 +1,7 @@
 ﻿public record DashboardIssueDto(
     string SystemId, // מזהה מערכת
     string SystemName, // שם מערכת
-    int RequiredMonths, // נדרש
-    int AllocatedMonths, // מוקצה
-    int Gap // פער
+    double RequiredMonths, // נדרש
+    double AllocatedMonths, // מוקצה
+    double Gap // פער
 );
