+++ practicum/BackEnd/HR System/DTOs/Dashboard/DashboardSummaryDto.cs
@@ -9,9 +9,9 @@ namespace HR_System.DTOs.Dashboard
         int ShortageSystemsCount,//מערכות במחסור
         int BalancedSystemsCount,//מערכות מאוזנות
         int ExcessSystemsCount,//מערכות בעודף
-        int TotalRequiredMonths,//חודשים נדרשים
-        int TotalAllocatedMonths,//חודשים  מוקצים
-        int TotalGap,//הפער הכולל
+        double TotalRequiredMonths,//חודשים נדרשים
+        double TotalAllocatedMonths,//חודשים  מוקצים
+        double TotalGap,//הפער הכולל
         int AvailableEmployeesCount,//עובדים פנויים 
         int BalancedEmployeesCount,//עובדים מאוזנים
         int OverloadedEmployeesCount,//עובדים עמוסים
