+++ practicum/BackEnd/HR System/DTOs/Employees/EmployeeAllocationEditDto.cs
@@ -5,7 +5,7 @@ namespace HR_System.DTOs.Employees
     {
         public string SystemId { get; init; } = null!;
         public string? RoleInSystem { get; init; }
-        public int? PlannedMonths { get; init; }
-        public int? ActualMonths { get; init; }
+        public double? PlannedMonths { get; init; }
+        public double? ActualMonths { get; init; }
     }
 }
