+++ practicum/BackEnd/HR System/DTOs/Systems/SystemCreateDto.cs
@@ -5,7 +5,7 @@ namespace HR_System.DTOs.Systems
         public string Name { get; set; } = null!;
         public string? OwnerManagerName { get; set; }
         public int Year { get; set; }
-        public int RequiredCapacityMonths { get; set; }
+        public double RequiredCapacityMonths { get; set; }
         public decimal AllocatedBudget { get; set; }
         public string? ManagementNote { get; set; }
         public bool IsActive { get; set; } = true;
